+++ src/hooks/useLocalStorage.ts
@@ -1,0 +1,42 @@
+import { useState } from 'react';
+
+
+export function useLocalStorage<T>(key: string, initialValue: T): [T, (value: T | ((val: T) => T)) => void] {
+    // Get from local storage then ...
+    // Parse stored json or if none return initialValue
+    const [storedValue, setStoredValue] = useState<T>(() => {
+        try {
+            const item = window.localStorage.getItem(key);
+            // Check if item is "undefined" string which can happen with JSON.stringify(undefined)
+            if (item === null || item === "undefined") {
+                return initialValue;
+            }
+            return JSON.parse(item);
+        } catch (error) {
+            console.log(error);
+            return initialValue;
+        }
+    });
+
+    // Return a wrapped version of useState's setter function that ...
+    // ... persists the new value to localStorage.
+    const setValue = (value: T | ((val: T) => T)) => {
+        try {
+            // Allow value to be a function so we have same API as useState
+            const valueToStore =
+                value instanceof Function ? value(storedValue) : value;
+
+            // Save state
+            setStoredValue(valueToStore);
+
+            // Save to local storage
+            if (typeof window !== "undefined") {
+                window.localStorage.setItem(key, JSON.stringify(valueToStore));
+            }
+        } catch (error) {
+            console.log(error);
+        }
+    };
+
+    return [storedValue, setValue];
+}